+++ lib/kettle/dev/changelog_cli.rb
@@ -740,13 +740,14 @@ module Kettle
       end
 
       def changelog_coverage_env
-        env = {
+        env = BundlerEnvGuard.unbundled_env.merge(
+          "RUBYOPT" => nil,
           "K_SOUP_COV_DO" => "true",
           "K_SOUP_COV_FORMATTERS" => "json",
           "K_SOUP_COV_MIN_HARD" => @enforce_coverage_thresholds ? "true" : "false",
           "K_SOUP_COV_MULTI_FORMATTERS" => "false",
           "K_SOUP_COV_OPEN_BIN" => ""
-        }
+        )
         gemfile = File.join(@coverage_root, "Gemfile")
         env["BUNDLE_GEMFILE"] = gemfile if File.file?(gemfile)
         env